% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capimatlabgen-package.r
\docType{package}
\name{capimatlabgen-package}
\alias{capimatlabgen}
\alias{capimatlabgen-package}
\title{A package to facilitate code generation to surface a C API}
\description{
A package to facilitate code generation to surface a C API
}
\details{
\tabular{ll}{
Package: \tab capimatlabgen\cr
Type: \tab Package\cr
Version: \tab 0.1-1 \cr
Date: \tab 2017-05-15 \cr
Release Notes: \tab Initial packaging for RPP and SWIFT matlab bindings.\cr
License: \tab GPL 2\cr
}
}
\author{
Jean-Michel Perraud \email{jean-michel.perraud_at_csiro.au}
}
\keyword{Matlab}

